feat: implement REST API

+++ backend/src/main/java/com/rayan/cinemaapi/service/BookingService.java
@@ -1,10 +1,12 @@
 package com.rayan.cinemaapi.service;
 
 import com.rayan.cinemaapi.entity.Booking;
+import com.rayan.cinemaapi.entity.Screening;
+import com.rayan.cinemaapi.entity.Seat;
+import com.rayan.cinemaapi.entity.User;
 import com.rayan.cinemaapi.exception.EntityNotFoundException;
 import com.rayan.cinemaapi.exception.InvalidBookingException;
 import com.rayan.cinemaapi.repository.BookingRepository;
-import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,9 +16,20 @@ import java.util.Objects;
 public class BookingService {
 
     private final BookingRepository bookingRepository;
+    private final ScreeningService screeningService;
+    private final UserService userService;
+    private final SeatService seatService;
 
-    public BookingService(BookingRepository bookingRepository) {
+    public BookingService(
+            BookingRepository bookingRepository,
+            ScreeningService screeningService,
+            UserService userService,
+            SeatService seatService
+    ) {
         this.bookingRepository = bookingRepository;
+        this.screeningService = screeningService;
+        this.userService = userService;
+        this.seatService = seatService;
     }
 
     public List<Booking> getBookings() {
@@ -30,31 +43,41 @@ public class BookingService {
     }
 
     public Booking createBooking(Booking booking) {
-        if (!Objects.equals(booking.getScreening().getRoom().getId(), booking.getSeat().getRoom().getId())) {
-            throw new InvalidBookingException();
-        }
+        resolveBooking(booking);
 
         return bookingRepository.save(booking);
     }
 
-    @Transactional
-    public Booking updateBooking(Booking booking) {
-        Booking existingBooking = getBooking(booking.getId());
+    public void deleteBooking(Long id) {
+        if (!bookingRepository.existsById(id)) {
+            throw new EntityNotFoundException("Booking", id);
+        }
+
+        bookingRepository.deleteById(id);
+    }
+
+    private void resolveBooking(Booking booking) {
+        Screening screening = screeningService.getScreening(
+                booking.getScreening().getId()
+        );
+
+        User user = userService.getUser(
+                booking.getUser().getId()
+        );
+
+        Seat seat = seatService.getSeat(
+                booking.getSeat().getSeatId()
+        );
 
         if (!Objects.equals(
-                booking.getScreening().getRoom().getId(),
-                booking.getSeat().getRoom().getId())) {
+                screening.getRoom().getId(),
+                seat.getRoom().getId()
+        )) {
             throw new InvalidBookingException();
         }
 
-        existingBooking.setUser(booking.getUser());
-        existingBooking.setScreening(booking.getScreening());
-        existingBooking.setSeat(booking.getSeat());
-
-        return existingBooking;
-    }
-
-    public void deleteBooking(Long id) {
-        bookingRepository.deleteById(id);
+        booking.setScreening(screening);
+        booking.setUser(user);
+        booking.setSeat(seat);
     }
 }
