+++ backend/src/main/java/com/rayan/cinemaapi/service/SeatService.java
@@ -2,7 +2,9 @@ package com.rayan.cinemaapi.service;
 
 import com.rayan.cinemaapi.entity.Seat;
 import com.rayan.cinemaapi.entity.SeatId;
+import com.rayan.cinemaapi.exception.EntityInUseException;
 import com.rayan.cinemaapi.exception.EntityNotFoundException;
+import com.rayan.cinemaapi.repository.BookingRepository;
 import com.rayan.cinemaapi.repository.SeatRepository;
 import org.springframework.stereotype.Service;
 
@@ -12,9 +14,11 @@ import java.util.List;
 public class SeatService {
 
     private final SeatRepository seatRepository;
+    private final BookingRepository bookingRepository;
 
-    public SeatService(SeatRepository seatRepository) {
+    public SeatService(SeatRepository seatRepository, BookingRepository bookingRepository) {
         this.seatRepository = seatRepository;
+        this.bookingRepository = bookingRepository;
     }
 
     public List<Seat> getSeats() {
@@ -32,6 +36,14 @@ public class SeatService {
     }
 
     public void deleteSeat(SeatId id) {
+        if (!seatRepository.existsById(id)) {
+            throw new EntityNotFoundException("Seat", id);
+        }
+
+        if (bookingRepository.existsBySeatId(id.getSeatLabel(), id.getRoomId())) {
+            throw new EntityInUseException("Seat", id);
+        }
+
         seatRepository.deleteById(id);
     }
 }
